+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Fixed
+  * AZ-1382: Fix `ssl_profile`block - Change type to list of object to enabled to allow the creation of mulitiple policy.
+
 # v7.7.1 - 2023-10-06
 
 Fixed